# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StoryVoice API
  description: |
    API for DawnoTemu/StoryVoice — user auth, voice cloning, story access, audio synthesis,
    and credits (Story Points) billing. This spec reflects the current server capabilities
    used by mobile apps, including credit charging, refunds, and status endpoints.
  version: 1.2.0
  contact:
    email: api@storyvoice.app
servers:
  - url: https://api.storyvoice.app
    description: Production server
  - url: http://localhost:8000
    description: Local development server

# Default security (empty). Protected operations specify bearerAuth explicitly.
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the /auth/login or /auth/refresh endpoints. Note that the user's email must be confirmed to access protected resources.
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        email_confirmed:
          type: boolean
          example: true
        is_active:
          type: boolean
          example: false
          description: Whether the user account is active (required for beta access)
        created_at:
          type: string
          format: date-time
          example: "2025-03-19T14:22:35.982Z"
        last_login:
          type: string
          format: date-time
          example: "2025-03-23T14:30:15.651Z"
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          description: New email address for the user.
        current_password:
          type: string
          description: Existing password required to confirm the profile change.
        new_password:
          type: string
          description: Replacement password for the account.
        new_password_confirm:
          type: string
          description: Confirmation of the replacement password.
      required:
        - current_password
    UpdateProfileResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
          example: Profile updated successfully.
        email_confirmation_required:
          type: boolean
          description: Indicates if the user must reconfirm their email address.
          example: true
        email_confirmation_error:
          type: string
          nullable: true
          description: Optional message if sending the confirmation email failed.
        password_updated:
          type: boolean
          description: Indicates if the password was changed during this update.
          example: false
    CreditLot:
      type: object
      properties:
        id:
          type: integer
        source:
          type: string
          example: monthly
        amount_granted:
          type: integer
          example: 20
        amount_remaining:
          type: integer
          example: 5
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_active:
          type: boolean
          description: True when the lot still has remaining credits and has not expired.
          example: true
    CreditTransactionEntry:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [credit, debit, refund, expire]
        amount:
          type: integer
        reason:
          type: string
          nullable: true
        status:
          type: string
        audio_story_id:
          type: integer
          nullable: true
        story_id:
          type: integer
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    CreditHistory:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreditTransactionEntry'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        next_offset:
          type: integer
          nullable: true
        applied_types:
          type: array
          items:
            type: string
            enum: [credit, debit, refund, expire]
    CreditSummaryResponse:
      type: object
      properties:
        balance:
          type: integer
        unit_label:
          type: string
        unit_size:
          type: integer
        lots:
          type: array
          items:
            $ref: '#/components/schemas/CreditLot'
        history:
          $ref: '#/components/schemas/CreditHistory'
        recent_transactions:
          type: array
          items:
            $ref: '#/components/schemas/CreditTransactionEntry'
    CreditHistoryResponse:
      $ref: '#/components/schemas/CreditHistory'
    DeleteAccountRequest:
      type: object
      additionalProperties: false
      properties:
        current_password:
          type: string
          description: Existing password required to confirm account deletion.
        reason:
          type: string
          description: Optional free-form reason provided by the user.
      required:
        - current_password
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Description of the error"
        details:
          type: string
          example: "Detailed error information"
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "secure_password"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOi..."
        refresh_token:
          type: string
          example: "eyJhbGciOi..."
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - password_confirm
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "secure_password"
        password_confirm:
          type: string
          format: password
          example: "secure_password"
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: "Registration successful. Please check your email to confirm your account."
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOi..."
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOi..."
    ResetPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
...
    NewPasswordRequest:
      type: object
      required:
        - new_password
        - new_password_confirm
      properties:
        new_password:
          type: string
          format: password
          example: "new_secure_password"
        new_password_confirm:
          type: string
          format: password
          example: "new_secure_password"
    NewPasswordResponse:
      type: object
      properties:
        message:
          type: string
          example: "Password has been reset successfully. You can now log in with your new password."
    ConfirmEmailResponse:
      type: object
      properties:
        message:
          type: string
          example: "Email confirmed successfully. You can now log in."
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request (missing fields or passwords don't match)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in a user
      description: Authenticate a user and receive JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account inactive or email not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Email not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the profile information of the currently authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - Authentication
      summary: Update current user profile
      description: Update the authenticated user's email and/or password after confirming the current password.
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          description: Invalid request (missing password confirmation, no changes, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Authentication
      summary: Delete current user account
      description: Deactivate and delete the authenticated user's account after confirming the current password. Cleanup runs asynchronously.
      operationId: deleteCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '202':
          description: Account deletion enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Missing current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error deleting the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/confirm-email/{token}:
    get:
      tags:
        - Authentication
      summary: Confirm email
      description: Confirm a user's email address using the token sent via email
      operationId: confirmEmail
      parameters:
        - name: token
          in: path
          description: Email confirmation token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmEmailResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/resend-confirmation:
    post:
      tags:
        - Authentication
      summary: Resend confirmation email
      operationId: resendConfirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email sent (if account exists)
        '400':
          description: Missing email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/reset-password-request:
    post:
      tags:
        - Authentication
      summary: Request password reset email
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Request processed
        '400':
          description: Missing email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/reset-password/{token}:
    post:
      tags:
        - Authentication
      summary: Reset password with token
      operationId: resetPasswordWithToken
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewPasswordResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /me/credits:
    get:
      tags:
        - Billing
      summary: Get current user's credit balance and history
      operationId: getMyCredits
      security:
        - bearerAuth: []
      parameters:
        - name: history_limit
          in: query
          description: Maximum number of recent transactions to include (default 20, max 100).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: history_offset
          in: query
          description: Offset into the transaction history for pagination.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: type
          in: query
          description: Comma-separated list of transaction types (credit, debit, refund, expire) to filter history.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Current balance, lots, and paginated history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditSummaryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account inactive or email not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/credits/history:
    get:
      tags:
        - Billing
      summary: List credit transactions with pagination
      operationId: getMyCreditHistory
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of transactions to return (default 20, max 100).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Offset into the transaction list.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: type
          in: query
          description: Comma-separated list of transaction types (credit, debit, refund, expire) to include.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of credit transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditHistoryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account inactive or email not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stories/{story_id}/credits:
    get:
      tags:
        - Billing
      summary: Estimate required credits for a story
      operationId: getStoryCredits
      parameters:
        - name: story_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Required credits for the story
          content:
            application/json:
              schema:
                type: object
                properties:
                  required_credits:
                    type: integer
                    example: 3
        '404':
          description: Story not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...